Move BreakerThread to its own file
