+++ src/v8_py_frontend/mini_racer.h
@@ -1,4 +1,3 @@
-#include <libplatform/libplatform.h>
 #include <v8.h>
 
 #include <map>
